[IMP] runbot_merge: inline stage_into, remove working copy
Turns out we can (most likely) do all the work in the local main repo
directly, so there is no need for the separate working copy, or the
exist stack, and thus no need for a separate function to be
context-managed.

Should have yagni'd and waited to setup the working copy stuff... oh
well.