[FIX] runbot: force update global_state
In some case the global state can be inconsistent, waiting when all
builds are done. My suspicion is that two build finish at the same time,
but the update is not done in database leading to an inconsistent state
if two build finishes at the same time.

This should solve this issue by forcing the update, leading to the
rollback of one of the transaction.