[IMP] runbot: improve has_pr button
Previously the has_pr button was a literal href which meant you couldn't
do both a search with the has_pr checked in a singular action.

Now the has_pr button acts as a checkbox (the real checkbox is hidden)
and basic support for keyboard controls has been added (tab, space and
enter).