[REF] runbot: rewrite stats page using owl
Removing very old javascript and using new tools will allow us to more
efficiently change and maintain the stat page in the futur.
This commit is meant to produce a 1:1 copy of the stats page like it was
before.
Further refactoring will happen to improve performance, layout and
component logic.