[REM] runbot: drop cron support
The runbot uses leader and builder processes for a while, mainly because
cron are not adapted for such work. Because of memory limit, lack of
controll of execution on multiple hosts, ... The cron was not tested for
a while since only the leader and builder scripts are used.

This commit removes the cron support from the runbot module for easier
maintenance.