[IMP] runbot: reference slot instead of builds
For upgrade, the reference builds are store as it is but we could store
slots instead.

The slot has a link to the build, but also gives additionnal
information, like the batch.

Moreover, it is possible to link a slot that has no build yet, which
could be usefull in the future for cross versions builds.