[IMP] runbot: add kanban view and stages to build error page
Introduce a kanban view to the runbot build error records.
Kanban cards show the most valuable field with relevant icons.

The concept of "state" was also introduced for the build error
model.
Currently there is 3 states:
 1. New: default state for any new error build
 2. Solved: when the issue should be solved
 3. Disabled: build error currently disabled (test-tags)

States are expected to be used by the user. For instance
after fixing an underteministic error the user can change the
state from new -> solved so that the user knows that he's actively
not working on it