[IMP] runbot: prevent saving on switching tab
Auto save mechanism when switching tab is unwanted in runbot, most of
the time we will use other tabs to get the data necessary for the record
we are modifying/creating and we don't care about saving yet.
We also prevent closing the window if the record is dirty.
Auto save in pager and context switch is still kept.