[IMP] runbot: allow to choose a docker image identifier
When one wants to update a Dockerfile with significant changes without
impacting regular builds, he needs to create a test Dockerfile
beforehand.

With this commit, when a Dockerfile is changed, the newly built image is
tagged with a `.future` suffix, that way, the old image is still used
until the image identifier is changed to the future one.

The `.future` image can be tested in separately before being
used in production.

Also, the previous image is tagged with a `.previous`suffix, preventing
it to be pruned.