[FIX] runbot: remove --user from pip
Experimentations are done to use venv inside the docker to avoid having
a strong dependency on the user creation to install pip.

The default --user option of pip is not compatible with the venv causing
error when the install really need to make some operation.

Looking at the commit that introduced it e1161a03e it wasn't really
usefull but mostly added as a good practice in case we wish to use the
same command on a non debian based system.