[IMP] runbot: add tracking on upgrade exception elements
Tracking the field as it can happen that we modify the elements instead
of creating a new upgrade when it is necessary in forward ports.
For example a new computed field on a mixin changed in stable, and the
mixin is implemented in more models in later versions.

Making a single exception makes it easier to track, but it is not known
at which point which elements have been added.