[IMP] runbot: uniquify with tree hash
It is quite frequent to have multiple builds with the same tree hash,
most likely when rewording a commit or stashing multiple ones. This can
lead to useless load since pushing squashed commit should have the same
result as pushing the original commit in terms of test (This is not the
case for config checking the diff, the commit messages ... the batch
dependant triggers)

This commit modifies the way the identifier is computed to use the
tree_hash instead of the commit_id. This way, we can link
builds with the same tree_hash together, and have faster results

Note that this commit still have a problem to solve, how to ensure to
send status on the new commits.