[IMP] runbot: add a button on trigger form to fill missing slots
When creating a new trigger, especially for nightly builds, it's
convenient to test the trigger. One way of doing that is by starting the
trigger manually on the desired bundle instead of waiting for a new
batch. But as the slot with the new trigger does not exists yet on
batches it's not possible.

With this commit, a button is added on the batch form in order to create
the missing slots in the last batches of the sticky bundles.