[IMP] runbot: per version frequency graphs
This commit proposes an alternate way to reprensent the 30 days history
Instead of using line height to represent the frequency, we use this
axis to represent versions. This allows to easily see if it appears more
in one version than another, but also if it starts to decrease/increase
in some version over time allowing to see the effect of forwardports.

While on it, improves the general speed of pages. Two compute methods
fetching a lot of data are replaced by targeted sql queries. The counts
are removed from the search pannel, as well as the "all links" tab from
the form view.