[FIX] runbot: fix restore on temporary dns failures
From times to times, it happens that a host has a temporary dns failure
causing the restore step to fail.

With this commit, we add the `--retry-on-host-error` to the wget command
which is designed for such a case. It will retry up to 20 times and will
increase the waiting time between each time by one second up to 10
seconds.