[IMP] runbot: notify all builds
part of #1127

One part of this commit is to ensure that we have a build on the slot
before calling github status, as well as simplifying the api.

The other parts sends the status in all cases, even if we link an
existing build. This will allow to force a status update on new commits
when a commit is linked to an existing build in #1127.

This may also have the effect to fix not sended status when an old build
is linked.