[IMP] runbot: add a commit link details page
A new frontend page and route is added to show details of commit_links.
The datails includes the diff with the base commit for the commit links
specified in the url.

Also, the commit links can be filtered by version and repository.
Version ids and repository ids have to be given in URL params as comma
separated ids.

The build error frequency graph is also improved. Each cell is now
clickable and links to the batches of the day. On that `batch by date`
page, a link is added to the commit details page, filtered by the
versions and repositories where the build error was seen.

The original diff page was a custom page made by @Xavier-Do

Further improvements have to be made:
- add a form on the batches by date page to customize the details
  filters
- improve the UI of the diff page