[IMP] runbot: network disabled by default
A first part made the network optionally disabled. This was tested in
the nightly builds and now that all related errors were fixed, we can
make it the default.

As a side effect, it will also now be impossible to add a requirement
without asking for a docker image to test it, which is a good thing.
It will also force to have one docker image per version of the
requirements, meaning that we won't have an image not coherent with the
debian control or requirements.