[IMP] runbot: use batch date for first seen date
The log date can be useful to identify a pattern in apparition
(alway arround midnight, ...) but it is usually less relevant for
apparition when when usually focus on the "commit" so the batch date

Moreover the log date can be misleading when the build ran in faketime.
Even if it is not always the case, it is possible to have log_dates in
the future.