[IMP] runbot: allow bundle tagging
The purppose of this commit is to facilitate the Odoo freeze process.

A new BundleTag model is added that allows to tag bundles.
A frontend controller is also added to list bundles with a particular
tag.

A "description" compute stored field is added on the bundle. By default
the PR titles are used to fill the description.

A "team_id" compute stored field is added to help classifying the
bundles. The computed attribution is based on the Odoo convention that a
branch name should end with the employee n-gram.