[FIX] runbot: forward X-Accel-Buffering header for streaming responses
Because of multi-layered use of nginx, responses with "X-Accel-Buffering: no"
is still buffered because the special header is absorbed by layer just above the
odoo app. As a result, it's impossible to test in runbot endpoints that stream
responses. To ensure that header is respected thoughout the stack, we propose in
this commit to pass the "X-Accel-Buffering" header.