[FIX] runbot: fix stats for dynamic steps
Since dynamic steps are a thing, it is possible to have the same step
multiple time in a build breaking the stats constraint.

Note that the step could be removed as it is not mandatory but having
doubled stats without a way to distinguish them is not ideal, adding the
dynamic step name for now.