[FIX] runbot: pagination padding in Standard mode
Since commit [^1], the addition of the HTML5 doctype made the browser's
rendering switch from "Quirks mode" to "Standard mode". This applies a
lot of subtle adjustments to the layout's handling and, in this case,
made the pagination's padding adjustment useless and even broken.

This commit removes the (now faulty) padding's override.

[^1]: odoo/runbot@dc6ffbda