[IMP] runbot: improve build error search
In some case we would like to have the error only breaking in some
versions or triggers.

This commit add two compued fields with search methods to add This
capability.

searching [('only_trigger_ids', 'in', [trigger_id])] will return errors
only breaging in trigger_id

searching
[('only_trigger_ids', 'any', [('display_name', 'like', 'Single')])]
will return errors only breaging in Single like triger (not in multi)

[('only_trigger_ids', 'any', [('display_name', 'like', 'Fake')])]
will return errors only breaging in Faketime triger

The same apples to versions but this is not working because of the search
methods traversing errors, contents and builds. We need to store the
version_ids but the migration takes 14 minutes, will be done in another pr.