[REM] runbot: remove cron logic
At the beginning of runbot the whole management of build was made during
cron.

Its was moves to dedicated services at some point to improve scalability
and reliability.

Right now, a cron would timeout or be killed  while fetching the repo or
doing heavy operations because of the workers limits, sometimes leaving
corrupted repositories behind.

This pr removes the cron related logic in order to cleanup the code and
officialy remove the support of the cron.