[REM] runbot: remove unused icp
Some icp where used for cron and are not used anymore.

One of them, runbot_do_fetch was still used but the value was false on
all odoo maintained instances.
The main reason is that when restarting, even if fetching all repos
will updated missed signals, it will also slow down the restart
potentially taking more time to reveal other issues with the new version.

This commits just remove those unused icp and the related code.