[FIX] runbot: fetch threehash
Before this commit, to export some threehash, the commit is used to
check if the commit is present or not.

This can be a problem when rebuilding an old build that was linked to a
previous one by a threehash after a forcepush, because the old commit
may not be known by github anymore.

Not sure why it was the case since it looks like fetch works fine with
a threehash, so fixing it by fetching the threehash should be enough.